// src/Models/OldEmandate/SearchData.ts
export interface Searchdetails {
    dataList: any;
    length: any;
    arr: string[];
    FromDate: string;
    Todate: string;
    bankdrop: string;
    ManDateID: string;
    DebtorAccountType: string;
    ConsumerReferenceNumber: string;
    MaximumAmount: string;
    InstructedAgentMemberName: string;
    DebtorName: string;
    ServiceProvider: string;
    DebtorAccountNumber: string;
    DebtorAgentID: string;
    Collection: string;
    Frequency: string;
    CreationDateTime: string;
}